requirements.txt
streamlit
supabase
python-dotenv
plotly
pandas
python-dateutil